Moving setup stuff to the main TestBase
